+++ tests/Tests/PHPUnit/Traits/CreatesApplicationTraitTest.php
@@ -39,6 +39,29 @@ class CreatesApplicationTraitTest extends AbstractTestCase
     }
 
     /**
+     * Test bootstrap file getter.
+     *
+     * @return void
+     */
+    public function testGetApplicationBootstrapFiles()
+    {
+        $list  = (array) $this->getApplicationBootstrapFiles();
+        $found = false;
+
+        $this->assertNotEmpty($list);
+
+        foreach ($list as $application_bootstrap_file) {
+            if (\file_exists($application_bootstrap_file)) {
+                $found = true;
+
+                break;
+            }
+        }
+
+        $this->assertTrue($found, 'Bootstrap file was not found');
+    }
+
+    /**
      * @param Application $app
      *
      * @throws \PHPUnit\Framework\ExpectationFailedException